requirements.txt
numpy
pandas
requests
flask
pytest
pygetwindow
pywinauto
pyautogui
pytesseract
Pillow
pywin32
opencv-python